// frontend/apps/thunder-develop/src/features/flows/models/base.ts
export interface FieldOption {
  /**
   * The key of the field option.
   */
  key: string;

  /**
   * The value of the field option.
   */
  value: string;

  /**
   * The label of the field option.
   */
  label: string;
}